scanner/requirements.txt
# Requirements for the scanner module of the application.

fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-multipart>=0.0.9
yara-python>=4.3.0
cachetools>=5.3.0 
clamd>=0.4.0